requirements.txt
aiohttp==3.9.1
aiohttp-cors==0.7.0
websockets==12.0
pandas==2.2.3
numpy>=1.26.2
asyncpg==0.29.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
PyJWT==2.8.0
bcrypt==4.1.2
ccxt>=4.0.0
ta>=0.11.0
# Ensure pip and setuptools are up to date for Railway
pip>=24.0
setuptools>=69.0.0
